+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: nfaust <nfaust@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/23 20:50:43 by nfaust            #+#    #+#             */
-/*   Updated: 2022/12/06 14:07:10 by nfaust           ###   ########.fr       */
+/*   Updated: 2024/01/26 15:40:20 by nfaust           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int		ft_putchar(char c);
 int		ft_formats(char c, va_list arg);
 int		ft_print_str(char *str);
 int		ft_puthex(unsigned int nbr, char *base, int check);
-int		ft_putptr(uintptr_t nbr, char *base, int check);
+int		ft_putptr(unsigned int nbr, char *base, int check);
 int		ft_printf(const char *str, ...);
 
 #endif
